requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.0.2
django-oauth-toolkit==1.2.0
djangorestframework==3.10.1
idna==2.8
oauthlib==3.0.2
Pillow==6.1.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
